feat(AuthRouter.ts, auth.test.ts): add user creation functionality to AuthRouter and tests
Add functionality to create a new user in the AuthRouter when a POST request is made to the SIGNUP_BASE_ROUTE. The user's name and password are required fields. If the user already exists, a ConflictException is thrown. If there is an error when creating the user, a BadRequestException is thrown.

Add tests to AuthRouter to test the new functionality. The tests use a mock of the UserRepository to simulate the database. The tests check that the correct response is returned when the user is created successfully, when the user already exists, and when there is an error creating the user.